Updated Fox Spawns
Both left and right spawn completed and do read all of my comments inside code
