+++ LOTS of CHICKS/Assets/Prefabs/Eggs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1cd1252bb8d9ca440860bde77a5318bc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
