+++ LOTS of CHICKS/Assets/Prefabs/Eggs/Egg.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 207450d225dbabc44b56eca0869876fa
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
