+++ LOTS of CHICKS/Assets/Prefabs/Fox/Fox left.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a89637fa767444144809ec2f20192ee9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
