+++ LOTS of CHICKS/Assets/Prefabs/Fox/Fox right.prefab.meta
