+++ LOTS of CHICKS/Assets/Scenes/Main.unity
@@ -410,6 +410,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 503922352}
+  - component: {fileID: 503922353}
   m_Layer: 0
   m_Name: Fox4 Spawner
   m_TagString: Untagged
@@ -432,6 +433,25 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &503922353
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 503922351}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2371c47a52031c3479b196ad0835ca50, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  foxPrefab: {fileID: 4466954169229106218, guid: a89637fa767444144809ec2f20192ee9, type: 3}
+  minSpawnTime: 3
+  maxSpawnTime: 7
+  spawnPositions:
+  - {fileID: 405695400}
+  - {fileID: 503922352}
+  - {fileID: 190109819}
 --- !u!1 &519420028
 GameObject:
   m_ObjectHideFlags: 0
