+++ LOTS of CHICKS/Assets/Scripts/Fox/FoxMovement_L2R.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1944f8c0b42120d47a9cab70a98bbc27
+guid: 9dbbc733614ea364d83e1173b7e78d48
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
